fix(usb): fix disk size overflow

+++ src/usb.rs
@@ -411,7 +411,7 @@ pub fn get_disk_free_space(path: &str) -> Option<u64> {
         unsafe {
             let mut stat: libc::statvfs = std::mem::zeroed();
             if statvfs(c_path.as_ptr(), &mut stat) == 0 {
-                Some(stat.f_bavail * stat.f_frsize as u64)
+                Some(stat.f_bavail as u64 * stat.f_frsize as u64)
             } else {
                 None
             }
@@ -427,7 +427,7 @@ pub fn get_disk_free_space(path: &str) -> Option<u64> {
         unsafe {
             let mut stat: libc::statvfs = std::mem::zeroed();
             if statvfs(c_path.as_ptr(), &mut stat) == 0 {
-                Some(stat.f_bavail * stat.f_frsize as u64)
+                Some(stat.f_bavail as u64 * stat.f_frsize as u64)
             } else {
                 None
             }
